replacing "Flag" class/object examples with a "Fruit"
I had to fix some Portuguese comments left behind and thought Fruit would be more fun as an example than Flag. Also there is the programming concept of a "boolean flag variable" that I don't want to bring to mind in the reader as it would be unrelated and confusing.